NEW: 툴 변경 기능, 캔버스 사이즈 미리보기 크기 동적 변경
