+++ src/components/molecules/ToolPreset/index.js
@@ -1,28 +1,25 @@
-// import ToolBox from "../../atoms/ToolBox";
 import { lazy, Suspense } from "react";
+import { useSelector } from "react-redux";
+
+import { selectToolPresets } from "../../../features/utility/utilitySlice";
 import MockToolBox from "../../atoms/ToolBox/MockToolBox";
 import styles from "./ToolPreset.module.scss";
 
 const ToolBox = lazy(() => import("../../atoms/ToolBox"));
 
 function ToolPreset() {
+  const tools = useSelector(selectToolPresets);
+
   return (
     <div className={styles.preset}>
-      <Suspense fallback={<MockToolBox />}>
-        <ToolBox tool="rectangle" />
-      </Suspense>
-      <Suspense fallback={<MockToolBox />}>
-        <ToolBox tool="ellipse" />
-      </Suspense>
-      <Suspense fallback={<MockToolBox />}>
-        <ToolBox tool="line" />
-      </Suspense>
-      <Suspense fallback={<MockToolBox />}>
-        <ToolBox tool="selector" />
-      </Suspense>
-      <Suspense fallback={<MockToolBox />}>
-        <ToolBox tool="text" />
-      </Suspense>
+      {tools.map(
+        (tool, i) =>
+          i < 5 && (
+            <Suspense fallback={<MockToolBox />} key={i}>
+              <ToolBox tool={tool} key={i} />
+            </Suspense>
+          )
+      )}
     </div>
   );
 }
