+++ src/components/scss/_colors.scss
@@ -16,3 +16,4 @@ $object-selected-tool-bg: #f4f7ff;
 $object-selected-tool-stroke: #1673ff;
 $object-selected-tool-stroke-bright: #c8deff;
 $object-hovered-bg: #e3eeff;
+$object-artboard-border: #aaaaaa;
