+++ src/features/globalStyles/globalStylesSlice.js
@@ -1,7 +1,6 @@
 import { createSlice } from "@reduxjs/toolkit";
 
 const initialState = {
-  tool: "selector",
   color: "#ffffff",
   thickness: 1,
   fontSize: 12,
@@ -11,9 +10,6 @@ const globalStylesSlice = createSlice({
   name: "globalStyles",
   initialState,
   reducers: {
-    setTool: (state, { payload }) => {
-      state.tool = payload;
-    },
     setColor: (state, { payload }) => {
       state.color = payload;
     },
@@ -26,8 +22,6 @@ const globalStylesSlice = createSlice({
   },
 });
 
-export const selectTool = (state) => state.workbench.present.globalStyles.tool;
-
 export const selectColor = (state) =>
   state.workbench.present.globalStyles.color;
 
@@ -37,7 +31,7 @@ export const selectThickness = (state) =>
 export const selectFontSize = (state) =>
   state.workbench.present.globalStyles.fontSize;
 
-export const { setColor, setFontSize, setThickness, setTool } =
+export const { setColor, setFontSize, setThickness } =
   globalStylesSlice.actions;
 
 export default globalStylesSlice.reducer;
