+++ src/features/utility/utilitySlice.js
@@ -3,10 +3,11 @@ import { createSlice } from "@reduxjs/toolkit";
 const initialState = {
   projectTitle: "untitled_project",
   isSelectorActivated: true,
-  tools: ["rectangle", "ellipse", "line", "seletor", "text"],
+  tools: ["rectangle", "ellipse", "line", "selector", "text", "canvas"],
   workingCanvas: {},
   selectedShapes: [],
   currentScale: 1,
+  currentTool: "selector",
 };
 
 const utilitySlice = createSlice({
@@ -25,6 +26,9 @@ const utilitySlice = createSlice({
     setProjectTitle: (state, { payload }) => {
       state.projectTitle = payload;
     },
+    setCurrentTool: (state, { payload }) => {
+      state.currentTool = payload;
+    },
   },
 });
 
@@ -34,11 +38,14 @@ export const selectCurrentScale = (state) => state.utility.currentScale;
 
 export const selectProjectTitle = (state) => state.utility.projectTitle;
 
+export const selectCurrentTool = (state) => state.utility.currentTool;
+
 export const {
   activateSelector,
   deactivateSelector,
   setCurrentScale,
   setProjectTitle,
+  setCurrentTool,
 } = utilitySlice.actions;
 
 export default utilitySlice.reducer;
