+++ src/styles.scss
@@ -9,5 +9,5 @@ body {
 }
 
 body {
-  background-color: #f8f8f8;
+  background-color: #e4e4e4;
 }
